+++ app/src/main/kotlin/catchup/app/service/bookmarks/OutlinedBookmark.kt
@@ -1,4 +1,4 @@
-package catchup.app.service
+package catchup.app.service.bookmarks
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.materialIcon
